Move checking cuda device check into function body.
Putting it in the function parameter appears to execute the code at time
of import which can lead to torch.CUDA complaining about multithreading
issues.

Fixes:

```
RuntimeError: Cannot re-initialize CUDA in forked subprocess. To use CUDA with multiprocessing, you must use the 'spawn' start method
```

when serving a flask app with gunicorn that imports this library.